==> studienarbeit/speeches/03_plotting.ipynb <==
import pandas as pd
import seaborn as sns
import json

from studienarbeit.utils.plots import Plots
from studienarbeit.utils.load import EDataTypes
from pathlib import Path
from dotenv import load_dotenv
from tqdm import tqdm
from nltk.tokenize import sent_tokenize

load_dotenv()
sns.set(style="white", palette="muted", rc={"figure.figsize": (20, 8)})
tqdm.pandas()

data_dir = Path("../../data/speeches")
data_type = EDataTypes.SPEECHES

with open("../../data/party_colors.json", "r", encoding="utf-8") as f:
    party_palette = json.load(f)
plots = Plots(data_type, data_dir, party_palette)

### Standard-Datensatz

df_speeches = pd.read_parquet(data_dir / "speeches.parquet")
df_speeches

with open("../../data/party_encoding.json", "r", encoding="utf-8") as f:
    party_encoding = json.load(f)
party_encoding = {v: k for k, v in party_encoding.items()}
df_speeches["party"] = df_speeches["party"].map(party_encoding)
df_speeches

plots.party_count(df_speeches)

df_speeches["char_count"] = df_speeches["clean_text"].progress_apply(len).astype("int16")
df_speeches["word_count"] = df_speeches["clean_text"].progress_apply(lambda x: len(x.split())).astype("int16")
df_speeches["sentence_count"] = (
    df_speeches["clean_text"].progress_apply(lambda x: len(sent_tokenize(x))).astype("int16")
)

plots.word_count(df_speeches, "word_count", "Anzahl an Wörtern nach Partei", 500, "Anzahl an Wörtern")

### Datensatz mit Sätzen als Einträgen

df_speeches_sentence = pd.read_parquet(data_dir / "speeches_sentence.parquet")
df_speeches_sentence

df_speeches_sentence["party"] = df_speeches_sentence["party"].map(party_encoding)
df_speeches_sentence

plots.party_count(df_speeches_sentence)

df_speeches_sentence["char_count"] = df_speeches_sentence["clean_text"].progress_apply(len).astype("int16")
df_speeches_sentence["word_count"] = (
    df_speeches_sentence["clean_text"].progress_apply(lambda x: len(x.split())).astype("int16")
)
df_speeches_sentence["sentence_count"] = (
    df_speeches_sentence["clean_text"].progress_apply(lambda x: len(sent_tokenize(x))).astype("int16")
)

plots.word_count(df_speeches_sentence, "word_count", "Anzahl an Wörtern nach Partei", 70, "Anzahl an Wörtern")

### Wordclouds (allgemein)

plots.wordclouds(df_speeches)
